chatbot-api/requirements.txt
fastapi
google-generativeai
faiss-cpu
numpy
